// src/app2/services/manage-email-template/manage-email-template.types.ts
export interface IEmailTemplate {
  id: number;
  name: string;
  content: string;
  firm_id: number;
  created_by: number;
  created_at: Date;
  updated_by?: any;
  updated_at?: any;
  is_active: boolean;
  is_default: boolean;
}

export interface IEmailTemplateRequest {
  name: string;
  content: string;
  id:number;
}

export interface IFirmPref {
  firm_id: number;
  created_by: number;
  created_at: Date;
  include_system: boolean;
  include_system_rating: boolean;
  hide_employees: boolean;
  add_generic_email: boolean;
  generic_email: string;
  add_sender_email: boolean;
  sender_email: string;
  auto_reminder: boolean;
  customize_intro: boolean;
  introduction: string;
  include_disclosure: boolean;
  disclosure: string;
  add_cc_email: boolean;
  cc_email: string;
  bcc_email: string;
  updated_by: number;
  updated_at: Date;
  regulatory_alert_frequency: number;
  password_expiration_frequency: number;
  password_reuse_limit: number;
  logo_link: string;
  disable_all_outbound_emails: boolean;
  disable_accept_decline_outbound_emails: boolean;
  set_materialADV_default?: any;
  set_firm_entity_default: string;
  ui_version: string;
  enable_questionnaire_numbering: boolean;
  enable_yaxis_entity: boolean;
  invert_color: boolean;
  set_preapproved_default: boolean;
  font_size: number;
  font_color: string;
  font_type: number;
  exclude_empty_response: boolean;
  exclude_comments: boolean;
  split_comments: boolean;
  include_internal_key: boolean;
  enable_grid_numbering: boolean;
  transpose_excel_columns: boolean;
  default_daterange_months: number;
  default_document_export_template_id?: any;
  default_email_template_message_id?: any;
  disallow_custom_email_template_edit: boolean;
  color_codes: string[];
  enable_category_instructions_export: boolean;
  enable_question_instructions_export: boolean;
  disable_response_followups: boolean;
  enable_system_document_tags: boolean;
  enable_system_strategies: boolean;
  enable_internal_notes_export: boolean;
  enable_followups_export: boolean;
  include_question_only_export: boolean;
  enable_notes_audit_export: boolean;
  enable_ratings_export: boolean;
  enableQASearch: boolean;
  enable_Review_Workflow: boolean;
  enable_track_changes: boolean;
  include_project_documents: boolean;
  review_workflow_mandatory_for_internal_diligence: boolean;
  review_workflow_mandatory_for_external_diligence: boolean;
  enable_internal_project_to_external_option: boolean;
  allow_review_template_in_standard_diligence: boolean;
  enable_rating_custom_fields_review: boolean;
  include_responses_excel_export: boolean;
  include_comments_excel_export: boolean;
  include_ratings_excel_export: boolean;
  enable_approval_flow_information_request: boolean;
  include_flags_excel_export: boolean;
  include_ratings_word_export: boolean;
  include_flags_word_export: boolean;
  enable_heatmap_dates: boolean;
  threshold_for_approval_flow: number;
  approver_list_for_information_request: any[];
  excel_bulk_import: boolean;
  default_response_word_limit?: any;
  accept_inbound_request: boolean;
  enable_inbound_module: boolean;
  enable_new_qa_look: boolean;
  doc_access_preference: string;
}